requirements.txt
Django==2.2.3
Pillow==6.1.0
PyMySQL==0.9.3
PyYAML==5.1.2
Whoosh==2.7.4
ajax==3.0.0
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-reversion==3.0.4
django-simple-captcha==0.5.11
docopt==0.6.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.13.1
idna==2.8
javascript==2015.32.59
jdcal==1.4.1
jieba==0.39
odfpy==1.4.0
openpyxl==2.6.3
pip==19.2.3
pipreqs==0.4.9
pysolr==3.8.1
pytz==2019.1
raven==6.10.0
redis==3.3.8
requests==2.22.0
setuptools==39.1.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
django-pure-pagination==0.3.0